cleanup, removing unnecessary props and such
GamerListing only cares about getting a gamerDetails object right now. It doesn't care about an index or action. It doesn't care about methods either. For example, removing a gamer would happen in BracketInfo because that's where the list of gamers is managed.

While it's possible (and sometimes cool) to have a GamerListing be able to remove itself, it's a lot hard to code, and you don't see it very often in the real world for that reason.